--- packages/editor/packages/editor-state/src/effects/exportWasm.ts
@@ -1,21 +1,0 @@
-import { EventDispatcher } from '../types';
-
-import type { State } from '../types';
-
-export default function exportWasm(state: State, events: EventDispatcher): void {
-	function onExportWasm() {
-		if (!state.callbacks.exportBinaryCode) {
-			console.warn('No exportProject callback provided');
-			return;
-		}
-
-		const projectName = 'project';
-		const fileName = `${projectName}.wasm`;
-
-		state.callbacks.exportBinaryCode(fileName).catch(error => {
-			console.error('Failed to export WebAssembly file:', error);
-		});
-	}
-
-	events.on('exportWasm', onExportWasm);
-}
