+++ packages/editor/packages/editor-state/src/features/project-export/__tests__/effect.test.ts
@@ -1,12 +1,11 @@
 import { describe, it, expect, beforeEach, vi, type MockInstance } from 'vitest';
 import createStateManager from '@8f4e/state-manager';
 
-import projectExport from './projectExport';
+import projectExport from '../effect';
+import { createMockState } from '../../../pureHelpers/testingUtils/testUtils';
+import { createMockEventDispatcherWithVitest } from '../../../pureHelpers/testingUtils/vitestTestUtils';
 
-import { createMockState } from '../pureHelpers/testingUtils/testUtils';
-import { createMockEventDispatcherWithVitest } from '../pureHelpers/testingUtils/vitestTestUtils';
-
-import type { State } from '../types';
+import type { State } from '../../../types';
 
 describe('projectExport', () => {
 	let mockState: State;
