+++ packages/editor/packages/editor-state/src/features/runtime/__tests__/effect.test.ts
@@ -1,10 +1,9 @@
 import { describe, it, vi, expect } from 'vitest';
 import createStateManager from '@8f4e/state-manager';
 
-import runtimeEffect from './runtime';
-
-import { createMockState } from '../pureHelpers/testingUtils/testUtils';
-import { createMockEventDispatcherWithVitest } from '../pureHelpers/testingUtils/vitestTestUtils';
+import runtimeEffect from '../effect';
+import { createMockState } from '../../../pureHelpers/testingUtils/testUtils';
+import { createMockEventDispatcherWithVitest } from '../../../pureHelpers/testingUtils/vitestTestUtils';
 
 describe('Runtime System', () => {
 	describe('Runtime lifecycle integration', () => {
