+++ packages/editor/packages/editor-state/src/index.ts
@@ -14,13 +14,12 @@ import configEffect from './features/config-compiler/effect';
 import contextMenu from './features/menu/effect';
 import graphicHelper from './effects/codeBlocks/graphicHelper';
 import editorSettings from './effects/editorSettings';
-import projectImport from './effects/projectImport';
-import projectExport from './effects/projectExport';
+import projectImport from './features/project-import/effect';
 import pianoKeyboard from './effects/codeBlocks/codeBlockDecorators/pianoKeyboard/interaction';
-import exportWasm from './effects/exportWasm';
+import projectExport from './features/project-export/effect';
 import viewport from './features/viewport/effect';
 import binaryAsset from './effects/binaryAssets';
-import runtime from './effects/runtime';
+import runtime from './features/runtime/effect';
 import blockTypeUpdater from './effects/codeBlocks/blockTypeUpdater';
 import shaderEffectsDeriver from './features/shader-effects/effect';
 import { validateFeatureFlags } from './pureHelpers/state/featureFlags';
@@ -74,7 +73,6 @@ export default function init(events: EventDispatcher, options: Options): StateMa
 	graphicHelper(store, events);
 	codeEditing(store, events);
 	projectExport(store, events);
-	exportWasm(state, events);
 	binaryAsset(state, events);
 	historyTracking(store, events);
 	events.dispatch('init');
